修复关闭原有服务脚本

+++ shell/stop_all.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
-jps |  grep 'blog' | awk '{print $1}' | xargs kill -15
+pids = jps |  grep 'blog' | awk '{print $1}'
+
+if [[ pids ]];then
+        jps |  grep 'blog' | awk '{print $1}' | xargs kill -15
+    else
+        echo "service is not running"
+fi
